+++ Tools/guiTools.py
@@ -10,6 +10,7 @@ from Tools.webhookTool import CheckForUpdates
 from screeninfo import get_monitors
 import Globals as gs
 from math import isnan
+from Tools.Items.Killer import DecodeNote
 
 
 # Create a window
@@ -253,36 +254,42 @@ def CreateTreeView(i_root, i_CodesData, i_RefreshInterval, i_RefreshCallback):
         tree.column('Code', width=0, stretch=tk.NO)                 # I'm going to keep Code hidden and add a new colum Notes
         tree.heading('Notes', text='Notes')
 
+        tree.column('File', width=240)
+        tree.column('Date', width=140)
+        #tree.column('Notes', width=490)
+        tree.column('Notes', width=390)
+        
+        tree.columnconfigure(0, weight=3)
+        tree.columnconfigure(1, weight=3)
+        tree.columnconfigure(3, weight=1)
+
         # Scrollbar
         scrollbar = Scrollbar(i_root, orient = 'vertical', command=tree.yview)
         tree.configure(yscrollcommand=scrollbar.set)
         scrollbar.pack(side='right', fill='y')
         tree.pack(side='left', fill='both', expand=True)
 
-        # Adjust column widths based on the data
-        def AdjustColumnSizes():
-            colIndex = {'File': 0, 'Date': 1, 'Notes': 3}
-            for col in colIndex:
-                maxWidth = max(
-                    len(str(item[colIndex[col]])) 
-                    if len(item) > colIndex[col] and item[colIndex[col]] is not None    # This way we cover for empty columns
-                    else 0 
-                    for item in i_CodesData
-                )
-                tree.column(col, width=maxWidth * 4)                                    # It doesn't really fit, so I'll multiply a bit
-
         # I need the window to refresh from time to time in case there's new data
         def FillTree():
+            gs.newCodeAdded = False
             existingItems = {tree.item(item, 'values') for item in tree.get_children()}     # Get all the items in the tree in a tuple, only the values
             sortedData = sorted(i_CodesData, key=lambda x: x[1], reverse=True)              # Sort by Date
-            if len(existingItems) == 0:                 # If there's no data in the table, just insert as is
+            if len(existingItems) == 0:                                                     # If there's no data in the table, just insert as is
                 for fileData in sortedData:
-                    if fileData not in existingItems and fileData[2]!='':                   # Only insert new values, avoid blank codes (this is for the future)
-                        tree.insert('', tk.END, values=fileData)
+                    noteDecoded = DecodeNote(fileData[3])                                   # Review what the code is to show something user readable
+                    row = list(fileData)
+                    row[3] = noteDecoded
+                    rowTuple = tuple(row)
+                    if rowTuple not in existingItems and fileData[2]!='':                   # Only insert new values, avoid blank codes (this is for the future)
+                        tree.insert('', tk.END, values=rowTuple)
             else:                                       # If there's data, I need to read from the bottom of sortedData, insert at the top of the tree
                 for fileData in reversed(sortedData):
-                    if fileData not in existingItems and fileData[2]!='':                   # Only insert new values, avoid blank codes (this is for the future)
-                        tree.insert('', 0, values=fileData)
+                    noteDecoded = DecodeNote(fileData[3])                                   # Review what the code is to show something user readable
+                    row = list(fileData)
+                    row[3] = noteDecoded
+                    rowTuple = tuple(row)
+                    if rowTuple not in existingItems and fileData[2]!='':                   # Only insert new values, avoid blank codes (this is for the future)
+                        tree.insert('', 0, values=rowTuple)
 
         # Event handler for double click, copies the code to the clipboard (I needed xclip on pop!_os for it to work, on windows there's no need in theory)
         def on_row_click(event):
@@ -302,7 +309,8 @@ def CreateTreeView(i_root, i_CodesData, i_RefreshInterval, i_RefreshCallback):
             if not gs.writingFlag:
                 nonlocal i_CodesData
                 i_CodesData = i_RefreshCallback()
-                FillTree()
+                if gs.newCodeAdded:
+                    FillTree()
                 i_root.after(i_RefreshInterval, refreshTree)
 
             if gs.configList['firstBoot']==True:        # Open the config window to get the path if this is the first boot
@@ -331,8 +339,6 @@ def CreateTreeView(i_root, i_CodesData, i_RefreshInterval, i_RefreshCallback):
         tree.bind('<Double-1>', on_row_click)           # Hook the double click event
         tree.bind('<Delete>', on_row_del)               # Hook for deleting rows
         FillTree()                                      # Fill the tree
-        if len(tree.get_children()) > 0:
-            AdjustColumnSizes()
         i_root.after(i_RefreshInterval, refreshTree)    # Hook the data refresh event
     except Exception as e:
         print(e)